Create a GUI with buttons taking up less space.
Reorganize IPA chart in a very strange way so that, the buttons are closer together. Some thought was put into finding ways to put graphemes with related meaning close together, while at the same time avoiding the traditional organization where it takes up too much space. The reason is that we don't want the person to have to scroll as much.